requirements.txt
numpy>=1.26
Pillow>=10.4
PyYAML>=6.0
protobuf>=5.27
sentencepiece>=0.2.0
torch>=2.4
transformers>=4.44
tqdm>=4.66
fastapi>=0.115
uvicorn>=0.32
